refactor: refactoring delete voluntary

+++ src/modules/voluntary/useCases/delete/DeleteVoluntaryController.ts
@@ -6,12 +6,13 @@ import { DeleteByIdVoluntaryUseCase } from './DeleteByIdVoluntaryUseCase';
 export class DeleteVoluntaryController {
   async handle(request: Request, response: Response): Promise<Response> {
     const { id } = request.params;
+    const userId = request.user.id;
 
     const deleteByIdVoluntaryUseCase = container.resolve(
       DeleteByIdVoluntaryUseCase,
     );
 
-    await deleteByIdVoluntaryUseCase.execute(id);
+    await deleteByIdVoluntaryUseCase.execute({ id, userId });
 
     return response
       .status(201)
